# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: UptivaLab API
  description: |
    Comprehensive uptime monitoring and status page platform API.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Features
    - HTTP/HTTPS monitoring
    - TCP port monitoring
    - Ping monitoring
    - DNS monitoring
    - Certificate expiry tracking
    - Synthetic browser testing
    - Docker container monitoring
    - gRPC service monitoring
    - Push-based heartbeat monitoring
    - Incident management
    - Maintenance windows
    - Public status pages
    - Multi-channel notifications
    
  version: 1.0.0
  contact:
    name: UptivaLab Support
    url: https://github.com/tekuonline/uptivaLab
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.uptivalab.com/api
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List notification channels
      description: Sensitive fields (passwords, API keys, tokens) are masked in the response
      responses:
        '200':
          description: List of notification channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationChannel'

    post:
      tags: [Notifications]
      summary: Create notification channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationChannelInput'
      responses:
        '200':
          description: Notification channel created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationChannel'
  /notifications/{id}:
    put:
      tags: [Notifications]
      summary: Update notification channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationChannelInput'
      responses:
        '200':
          description: Notification channel updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationChannel'

    delete:
      tags: [Notifications]
      summary: Delete notification channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification channel deleted
  /notifications/test:
    post:
      tags: [Notifications]
      summary: Test notification channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channelId]
              properties:
                channelId:
                  type: string
      responses:
        '200':
          description: Test notification sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string

  # Status Pages
...
  /settings/api-keys:
    get:
      tags: [Settings]
      summary: List API keys
      description: Returns only metadata, tokens are never exposed after creation
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'

    post:
      tags: [Settings]
      summary: Create API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label:
                  type: string
      responses:
        '200':
          description: API key created (token only shown once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  label:
                    type: string
                  token:
                    type: string
                    description: This is the only time the token will be shown
  /settings/api-keys/{id}:
    delete:
      tags: [Settings]
      summary: Delete API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API key deleted
  /settings/docker-hosts:
    get:
      tags: [Settings]
      summary: List Docker hosts
      responses:
        '200':
          description: List of Docker hosts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DockerHost'

    post:
      tags: [Settings]
      summary: Add Docker host
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, socketPath]
              properties:
                name:
                  type: string
                socketPath:
                  type: string
      responses:
        '200':
          description: Docker host added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DockerHost'
  /settings/remote-browsers:
    get:
      tags: [Settings]
      summary: List remote browsers
      responses:
        '200':
          description: List of remote browsers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RemoteBrowser'

    post:
      tags: [Settings]
      summary: Add remote browser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, url]
              properties:
                name:
                  type: string
                url:
                  type: string
                  format: uri
      responses:
        '200':
          description: Remote browser added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoteBrowser'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    NotificationChannel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [email, slack, discord, telegram, webhook, apprise]
        config:
          type: object
          description: Sensitive fields are masked with ***MASKED***
          additionalProperties: true
        enabled:
          type: boolean
    NotificationChannelInput:
      type: object
      required: [name, type, config]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [email, slack, discord, telegram, webhook, apprise]
        config:
          type: object
          description: Configuration varies by notification type
          additionalProperties: true
        enabled:
          type: boolean
          default: true
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        lastUsedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    DockerHost:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        socketPath:
          type: string
    RemoteBrowser:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
          format: uri
...